--- components/ColumnComponents/index.js
@@ -1,1 +1,0 @@
-export { default as PercentContainer } from './PercentContainer';
