+++ components/PercentContainer/PercentContainer.jsx
@@ -2,9 +2,8 @@ import React from 'react';
 import LinearProgress from 'material-ui/LinearProgress';
 // import PiePercent from '../../PiePercent';
 import styles from './PercentContainer.css';
-import { getPercentWin } from '../../../utility';
+import { getPercentWin } from '../../utility';
 // const getPercentWin = (wins, games) => (games ? Math.ceil(1000 * (wins / games)) / 10 : 0);
-
 
 // the LinearProgress can be swapped for PiePercent if it turns out people prefer that
 export default ({ wins, games }) => (
