+++ components/PercentContainer/index.js
@@ -1,0 +1,3 @@
+import PercentContainer from './PercentContainer';
+
+export default PercentContainer;
