+++ components/Table/TableHeaderColumn.jsx
@@ -5,7 +5,6 @@ import {
   getSortIcon,
 } from './tableHelpers';
 import styles from './Table.css';
-import { Text } from '../Text';
 
 export default ({ column, sortClick, sortField, sortState, totalWidth }) => (
   <MaterialTableHeaderColumn
@@ -15,9 +14,10 @@ export default ({ column, sortClick, sortField, sortState, totalWidth }) => (
       className={column.sortFn ? styles.headerCell : styles.headerCellNoSort}
       onClick={() => column.sortFn && sortClick(column.field, sortState, column.sortFn)}
     >
-      <Text size={14}>{column.displayName}
+      <div>
+      {column.displayName}
       {column.sortFn && getSortIcon(sortState, sortField, column.field, { height: 14, width: 14 })}
-      </Text>
+      </div>
     </div>
   </MaterialTableHeaderColumn>
 );
