--- components/Table/columnDefinitions/index.js
@@ -1,5 +1,0 @@
-export { default as playerMatchesColumns } from './playerMatchesColumns';
-export { default as playerPeersColumns } from './playerPeersColumns';
-export { default as playerHeroesColumns, playerHeroesOverviewColumns } from './playerHeroesColumns';
-export { default as playerRecordsColumns } from './playerRecordsColumns';
-export { overviewColumns } from './matchColumns';
