--- components/Text/Text.css
@@ -1,1 +1,0 @@
-@import "../palette.css";
