--- components/Text/index.js
@@ -1,3 +1,0 @@
-import Text from './Text';
-
-export { Text };
