+++ package.json
@@ -59,11 +59,10 @@
     "expose-loader": "^0.7.1",
     "file-loader": "^0.9.0",
     "json-loader": "^0.5.4",
-    "npm-check-updates": "2.8.0",
     "postcss-browser-reporter": "^0.5.0",
     "postcss-color-function": "^2.0.1",
     "postcss-cssnext": "^2.5.2",
-    "postcss-import": "8.0.2",
+    "postcss-import": "8.1.0",
     "postcss-loader": "^0.11.0",
     "postcss-reporter": "^1.3.3",
     "style-loader": "^0.13.1",
