+++ webpack.config.js
@@ -8,7 +8,7 @@ const isProd = process.env.NODE_ENV === 'production';
 const config = {
   entry:
   {
-    index: './index.js',
+    index: './index.jsx',
   },
   output:
   {
